Choose intelligently the prototype name to avoid conflicts
Because of the DIR.PROTO notation, it is possible to have two prototypes
with the same shortname while their longname are different. This patch
makes PROTOTYPE.name intelligent to be unique among all prototypes. The
drawback is that the prototype name can change to include directories if
another PROTOTYPE is added and have a conflicting name.

But the error mesages will show exactly which prototype is concerned
(provided the conflicting prototypes are all used).